Ising_DMRG/pyproject.toml
[project]
name = "ising-dmrg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.5",
    "physics-tenpy>=1.0.6",
]
